backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Redis and Celery
redis==4.6.0
celery[redis]==5.3.4

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Authentication and security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
typer==0.9.0
rich==13.7.0
loguru==0.7.2

# Data processing
pandas==2.1.4
numpy==1.26.2

# File handling
chardet==5.2.0
xml-python==0.4.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Archive handling
rarfile==4.1
py7zr==0.20.2

# System monitoring
psutil==5.9.6

# Monitoring
prometheus-client==0.19.0

# Image processing
Pillow==10.1.0

# Validation
email-validator==2.1.0